refactor: Rename `narwhals.dtypes.classes` -> `narwhals.dtypes._classes`
1/2 in https://github.com/narwhals-dev/narwhals/pull/3396#discussion_r2678724894
